# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TA Assistant API",
    "summary": "Market Regime + Ticker Analysis for the Next.js frontend.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/screener/status": {
      "get": {
        "tags": [
          "screener"
        ],
        "summary": "Status",
        "operationId": "status_api_screener_status_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Status Api Screener Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/screener/candidates": {
      "get": {
        "tags": [
          "screener"
        ],
        "summary": "Candidates",
        "operationId": "candidates_api_screener_candidates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScreenerCandidate"
                  },
                  "type": "array",
                  "title": "Response Candidates Api Screener Candidates Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Add Candidate",
        "description": "Manually add a ticker: validate + score it like a scanned candidate.",
        "operationId": "add_candidate_api_screener_candidates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "screener"
        ],
        "summary": "Clear Candidates",
        "operationId": "clear_candidates_api_screener_candidates_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "type": "object",
                  "title": "Response Clear Candidates Api Screener Candidates Delete"
                }
              }
            }
          }
        }
      }
    },
    "/api/screener/candidates/{symbol}": {
      "delete": {
        "tags": [
          "screener"
        ],
        "summary": "Delete Candidate",
        "operationId": "delete_candidate_api_screener_candidates__symbol__delete",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Candidate Api Screener Candidates  Symbol  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screener/alpha/evaluate": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Evaluate",
        "operationId": "evaluate_api_screener_alpha_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Evaluate Api Screener Alpha Evaluate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/screener/alpha/{symbol}": {
      "delete": {
        "tags": [
          "screener"
        ],
        "summary": "Delete Alpha",
        "operationId": "delete_alpha_api_screener_alpha__symbol__delete",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Alpha Api Screener Alpha  Symbol  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screener/alpha/refresh": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_screener_alpha_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Refresh Api Screener Alpha Refresh Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "AddRequest"
      },
...
      "AlphaReason": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "status": {
            "$ref": "#/components/schemas/Bias",
            "default": "neutral"
          }
        },
        "type": "object",
        "required": [
          "category",
          "detail"
        ],
        "title": "AlphaReason",
        "description": "One structured, reviewable reason behind an alpha verdict."
      },
      "AlphaVerdict": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "is_alpha": {
            "type": "boolean",
            "title": "Is Alpha",
            "default": false
          },
          "conviction": {
            "type": "integer",
            "title": "Conviction",
            "default": 0
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "default": ""
          },
          "sector": {
            "type": "string",
            "title": "Sector",
            "default": ""
          },
          "regime_alignment": {
            "type": "string",
            "title": "Regime Alignment",
            "default": ""
          },
          "entry": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry"
          },
          "stop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "rr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rr"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/AlphaReason"
            },
            "type": "array",
            "title": "Reasons"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "deterministic"
          },
          "inputs": {
            "type": "string",
            "title": "Inputs",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "AlphaVerdict",
        "description": "agent2/agent3 output: is this a genuine alpha opportunity, why, and what it saw."
      },
...
      "Bias": {
        "type": "string",
        "enum": [
          "bullish",
          "neutral",
          "bearish"
        ],
        "title": "Bias"
      },
      "CandidateStatus": {
        "type": "string",
        "enum": [
          "pending",
          "evaluating",
          "alpha",
          "not_alpha"
        ],
        "title": "CandidateStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScreenFactor": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "status": {
            "$ref": "#/components/schemas/Bias",
            "default": "neutral"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "ScreenFactor",
        "description": "One pre-screen reading shown on a scanner candidate row."
      },
      "ScreenerCandidate": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          },
          "status": {
            "$ref": "#/components/schemas/CandidateStatus",
            "default": "pending"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/ScreenFactor"
            },
            "type": "array",
            "title": "Factors"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "screen"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlphaVerdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "sector",
          "generated_at"
        ],
        "title": "ScreenerCandidate",
        "description": "A scanner hit \u2014 a ticker that MIGHT have a good trade (cheap pre-screen, not yet\nthe full agent analysis). The user selects these to push into the alpha pipeline.\nOnce evaluated, it carries the agent's verdict (reasons + the inputs it saw)."
      },
...
      "SymbolsRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols",
            "default": []
          }
        },
        "type": "object",
        "title": "SymbolsRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}